requirements.txt
certifi==2021.10.8
chardet==4.0.0
idna==3.3
requests==2.26.0
urllib3==1.26.7
cffi==1.15.0
cryptography==35.0.0
decorator==5.1.0
future==0.18.2
mccabe==0.6.1
pycparser==2.20
python-telegram-bot==13.7
schedule==1.1.0
six==1.16.0
snowballstemmer==2.1.0
tornado==6.1